fix broken migration
A recent change in the auth.identities table made the following changes:

1. The id column was renamed to provider_id.
2. A new id column which should be a random uuid.

This change broke our "fix_auth_schema_values" migration. To fix it we updated the insert statement to include email as the provider id value and a random uuid as the id. Since for email provider the provider_id should be email id in lowercase we added a new migration "update_provider_id" to update the provider_id column to the email column value.

Updating an older migration is bad practice but I don't think we have any other option now.